@@ -1,5 +1,6 @@
-module github.com/sijoma/console-customer-api-go
+module github.com/GIT_USER_ID/GIT_REPO_ID
 
-go 1.13
+go 1.18
 
-require golang.org/x/oauth2 v0.0.0-20210323180902-22b0adad7558
+require (
+)
