@@ -1,7 +1,7 @@
 /*
-Camunda Cloud Management API
+Camunda Management API
 
-Manage Camunda Cloud Clusters and API Clients via API.
+Manage Camunda Clusters and API Clients via API.
 
 API version: 1.3.3
 */
@@ -13,6 +13,9 @@ package openapi
 import (
 	"encoding/json"
 )
+
+// checks if the ClusterChannel type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &ClusterChannel{}
 
 // ClusterChannel The software channel your cluster is running on.
 type ClusterChannel struct {
@@ -52,7 +55,7 @@ func (o *ClusterChannel) GetName() string {
 // GetNameOk returns a tuple with the Name field value
 // and a boolean to check if the value has been set.
 func (o *ClusterChannel) GetNameOk() (*string, bool) {
-	if o == nil  {
+	if o == nil {
 		return nil, false
 	}
 	return &o.Name, true
@@ -76,7 +79,7 @@ func (o *ClusterChannel) GetUuid() string {
 // GetUuidOk returns a tuple with the Uuid field value
 // and a boolean to check if the value has been set.
 func (o *ClusterChannel) GetUuidOk() (*string, bool) {
-	if o == nil  {
+	if o == nil {
 		return nil, false
 	}
 	return &o.Uuid, true
@@ -88,14 +91,18 @@ func (o *ClusterChannel) SetUuid(v string) {
 }
 
 func (o ClusterChannel) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if true {
-		toSerialize["name"] = o.Name
-	}
-	if true {
-		toSerialize["uuid"] = o.Uuid
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o ClusterChannel) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	toSerialize["name"] = o.Name
+	toSerialize["uuid"] = o.Uuid
+	return toSerialize, nil
 }
 
 type NullableClusterChannel struct {
